Updated solution code for Words app
