--- app/src/main/java/com/example/wordsapp/DetailListFragment.kt
@@ -1,88 +1,0 @@
-/*
- * Copyright (C) 2020 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.example.wordsapp
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.DividerItemDecoration
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.wordsapp.databinding.FragmentDetailListBinding
-
-/**
- * Displays a [RecyclerView] of words with search buttons to look them up.
- */
-class DetailListFragment : Fragment() {
-    private lateinit var letterId: String
-
-    private var _binding: FragmentDetailListBinding? = null
-
-    // This property is only valid between onCreateView and
-    // onDestroyView.
-    private val binding get() = _binding!!
-
-
-    /**
-     * Provides global access to these variables from anywhere in the app
-     * via DetailListFragment.<variable> without needing to create
-     * a DetailListFragment instance.
-     */
-    companion object {
-        val LETTER = "letter"
-        val SEARCH_PREFIX = "https://www.google.com/search?q="
-    }
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        // Retrieve the LETTER from the Fragment arguments
-        arguments?.let {
-            letterId = it.getString(LETTER).toString()
-        }
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Retrieve and inflate the layout for this fragment
-        _binding = FragmentDetailListBinding.inflate(inflater, container, false)
-        val view = binding.root
-        return view
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        val recyclerView = binding.recyclerView
-        recyclerView.layoutManager = LinearLayoutManager(requireContext())
-        recyclerView.adapter = WordAdapter(letterId, requireContext())
-
-        // Adds a [DividerItemDecoration] between items
-        recyclerView.addItemDecoration(
-            DividerItemDecoration(context, DividerItemDecoration.VERTICAL)
-        )
-    }
-
-    /**
-     * Frees the binding object when the Fragment is destroyed.
-     */
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
