+++ app/src/main/java/com/example/wordsapp/LetterAdapter.kt
@@ -15,6 +15,7 @@
  */
 package com.example.wordsapp
 
+import android.content.Intent
 import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
@@ -24,6 +25,7 @@ import android.widget.Button
 import androidx.annotation.RequiresApi
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
+
 
 /**
  * Adapter for the [RecyclerView] in [MainActivity].
@@ -52,7 +54,7 @@ class LetterAdapter :
         val layout = LayoutInflater
             .from(parent.context)
             .inflate(R.layout.item_view, parent, false)
-
+        
         // Setup custom accessibility delegate to set the text read
         layout.accessibilityDelegate = Accessibility
         return LetterViewHolder(layout)
@@ -67,11 +69,7 @@ class LetterAdapter :
 
         // Assigns a [OnClickListener] to the button contained in the [ViewHolder]
         holder.button.setOnClickListener {
-            // Create an action from WordList to DetailList
-            // using the required arguments
-            val action = WordListFragmentDirections
-                .actionWordListToDetailList(letter = holder.button.text.toString())
-            // Navigate using that action
+            val action = LetterListFragmentDirections.actionLetterListFragmentToWordListFragment(letter = holder.button.text.toString())
             holder.view.findNavController().navigate(action)
         }
     }
