+++ app/src/main/java/com/example/wordsapp/LetterListFragment.kt
@@ -1,0 +1,91 @@
+package com.example.wordsapp
+
+import android.os.Bundle
+import android.view.*
+import androidx.core.content.ContextCompat
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.GridLayoutManager
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.example.wordsapp.databinding.FragmentLetterListBinding
+
+class LetterListFragment : Fragment() {
+
+    private var _binding: FragmentLetterListBinding? = null
+
+    private val binding get() = _binding!!
+
+    private lateinit var recyclerView: RecyclerView
+
+    private var isLinearLayoutManager = true
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setHasOptionsMenu(true)
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentLetterListBinding.inflate(inflater, container, false)
+        val view = binding.root
+        return view
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        recyclerView = binding.recyclerView
+        chooseLayout()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.layout_menu, menu)
+
+        val layoutButton = menu.findItem(R.id.action_switch_layout)
+        setIcon(layoutButton)
+    }
+
+    private fun chooseLayout() {
+        when (isLinearLayoutManager) {
+            true -> {
+                recyclerView.layoutManager = LinearLayoutManager(context)
+                recyclerView.adapter = LetterAdapter()
+            }
+            false -> {
+                recyclerView.layoutManager = GridLayoutManager(context, 4)
+                recyclerView.adapter = LetterAdapter()
+            }
+        }
+    }
+
+    private fun setIcon(menuItem: MenuItem?) {
+        if (menuItem == null)
+            return
+
+        menuItem.icon =
+            if (isLinearLayoutManager)
+                ContextCompat.getDrawable(this.requireContext(), R.drawable.ic_grid_layout)
+            else ContextCompat.getDrawable(this.requireContext(), R.drawable.ic_linear_layout)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            R.id.action_switch_layout -> {
+                isLinearLayoutManager = !isLinearLayoutManager
+                chooseLayout()
+                setIcon(item)
+
+                return true
+            }
+
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
+
+}
