+++ app/src/main/java/com/example/wordsapp/WordAdapter.kt
@@ -28,7 +28,7 @@ import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
 
 /**
- * Adapter for the [RecyclerView] in [DetailListFragment].
+ * Adapter for the [RecyclerView] in [DetailActivity].
  */
 class WordAdapter(private val letterId: String, context: Context) :
     RecyclerView.Adapter<WordAdapter.WordViewHolder>() {
@@ -68,6 +68,7 @@ class WordAdapter(private val letterId: String, context: Context) :
 
         // Setup custom accessibility delegate to set the text read
         layout.accessibilityDelegate = Accessibility
+
         return WordViewHolder(layout)
     }
 
@@ -85,7 +86,7 @@ class WordAdapter(private val letterId: String, context: Context) :
 
         // Assigns a [OnClickListener] to the button contained in the [ViewHolder]
         holder.button.setOnClickListener {
-            val queryUrl: Uri = Uri.parse("${DetailListFragment.SEARCH_PREFIX}${item}")
+            val queryUrl: Uri = Uri.parse("${WordListFragment.SEARCH_PREFIX}${item}")
             val intent = Intent(Intent.ACTION_VIEW, queryUrl)
             context.startActivity(intent)
         }
