+++ app/src/main/res/layout/fragment_letter_list.xml
@@ -10,7 +10,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:clipToPadding="false"
-        android:padding="16dp"
-        tools:listitem="@layout/item_view" />
+        android:padding="16dp" />
 
 </FrameLayout>
