+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,39 +1,23 @@
-<?xml version="1.0" encoding="utf-8"?><!--
-     Copyright (C) 2020 The Android Open Source Project
-
-     Licensed under the Apache License, Version 2.0 (the "License");
-     you may not use this file except in compliance with the License.
-     You may obtain a copy of the License at
-
-          http://www.apache.org/licenses/LICENSE-2.0
-
-     Unless required by applicable law or agreed to in writing, software
-     distributed under the License is distributed on an "AS IS" BASIS,
-     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-     See the License for the specific language governing permissions and
-     limitations under the License.
--->
+<?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/wordList">
+    app:startDestination="@id/letterListFragment">
 
     <fragment
-        android:id="@+id/wordList"
-        android:label="@string/app_name"
-        android:name="com.example.wordsapp.WordListFragment"
-        tools:layout="@layout/fragment_word_list">
+        android:id="@+id/letterListFragment"
+        android:name="com.example.wordsapp.LetterListFragment"
+        android:label="@string/app_name">
         <action
-            android:id="@+id/action_wordList_to_detailList"
-            app:destination="@id/detailList"
-            app:popUpToInclusive="true" />
+            android:id="@+id/action_letterListFragment_to_wordListFragment"
+            app:destination="@id/wordListFragment" />
     </fragment>
     <fragment
-        android:id="@+id/detailList"
-        android:name="com.example.wordsapp.DetailListFragment"
-        android:label="@string/fragment_label"
-        tools:layout="@layout/fragment_detail_list">
+        android:id="@+id/wordListFragment"
+        android:name="com.example.wordsapp.WordListFragment"
+        android:label="@string/word_list_fragment_label"
+        tools:layout="@layout/fragment_word_list">
         <argument
             android:name="letter"
             app:argType="string" />
